@@ -4,16 +4,16 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.1.4 <4.0.0'
+  sdk: '>=3.6.0 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
   yaru: ^1.1.0
   pine: ^1.0.2
-  auto_route: ^7.8.4
+  auto_route: ^9.0.0
   flutter_bloc: ^8.1.3
-  freezed_annotation: ^2.4.1
+  freezed_annotation: ^3.0.0
   hydrated_bloc: ^9.1.2
   intl: any
   path_provider: ^2.1.1
@@ -24,7 +24,7 @@ dependencies:
   talker: ^3.1.4
   reactive_forms: ^16.1.1
   yaru_widgets: ^3.2.2
-  file_picker: ^6.0.0
+  file_picker: ^8.0.0
   flutter_staggered_grid_view: ^0.7.0
   path: ^1.8.3
 
@@ -33,9 +33,9 @@ dev_dependencies:
     sdk: flutter
   flutter_lints: ^3.0.0
   bloc_test: ^9.1.5
-  auto_route_generator: ^7.3.2
+  auto_route_generator:
   build_runner: ^2.4.6
-  freezed: ^2.4.5
+  freezed:
   mockito: ^5.4.2
   mocktail: ^1.0.1
   data_fixture_dart: ^3.0.0
