@@ -3,11 +3,11 @@ import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_essentials_kit/flutter_essentials_kit.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_staggered_grid_view/flutter_staggered_grid_view.dart';
 import 'package:tlp_ui/blocs/configuration/configuration_bloc.dart';
 import 'package:tlp_ui/features/routing/app_router.dart';
 import 'package:tlp_ui/models/tlp_configuration/tlp_configuration.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:tlp_ui/widgets/file_card.dart';
 import 'package:tlp_ui/widgets/or_separator.dart';
 import 'package:yaru_widgets/yaru_widgets.dart';
@@ -40,97 +40,104 @@ class _WelcomePageState extends State<WelcomePage> {
   @override
   Widget build(BuildContext context) => Scaffold(
         body: BlocConsumer<ConfigurationBloc, ConfigurationState>(
-          listener: (context, state) => state.whenOrNull(
-            loaded: (configuration) => _navigateToMain(context, configuration),
-          ),
-          buildWhen: (_, state) => state.maybeWhen(
-            initial: () => true,
-            fetching: () => true,
-            fetched: (_) => true,
-            orElse: () => false,
-          ),
+          listener: (context, state) => switch (state) {
+            LoadedConfigurationState(:final configuration) =>
+              _navigateToMain(context, configuration),
+            _ => null,
+          },
+          buildWhen: (_, state) => switch (state) {
+            InitialConfigurationState() ||
+            FetchingConfigurationState() ||
+            FetchedConfigurationState() =>
+              true,
+            _ => false,
+          },
           builder: (context, state) => Container(
             alignment: Alignment.center,
             padding: const EdgeInsets.all(64.0),
-            child: state.maybeWhen(
-              fetching: () => const CircularProgressIndicator(),
-              orElse: () => IntrinsicWidth(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.stretch,
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    Padding(
-                      padding: const EdgeInsets.symmetric(vertical: 16.0),
-                      child: Text(
-                        AppLocalizations.of(context)?.label_welcome ??
-                            'label_welcome',
-                        style: Theme.of(context).textTheme.displayMedium,
-                        textAlign: TextAlign.center,
-                      ),
-                    ),
-                    state.maybeWhen(
-                      fetched: (files) => Column(
-                        crossAxisAlignment: CrossAxisAlignment.stretch,
-                        children: [
-                          YaruSection(
-                            margin: const EdgeInsets.symmetric(vertical: 16.0),
-                            padding: const EdgeInsets.all(16.0),
-                            headline: Text(AppLocalizations.of(context)
-                                ?.label_select_configuration ??
-                                'label_select_configuration'),
-                            child: SizedBox(
-                              height: 500.0,
-                              width: 1000.0,
-                              child: AlignedGridView.extent(
-                                padding: const EdgeInsets.all(16.0),
-                                maxCrossAxisExtent: 200.0,
-                                mainAxisSpacing: 16.0,
-                                crossAxisSpacing: 16.0,
-                                itemBuilder: (context, index) => FileCard(
-                                  files[index],
-                                  selected: files[index] == _selectedPath,
-                                  onTap: () => setState(
-                                        () => _selectedPath = files[index],
-                                  ),
-                                ),
-                                itemCount: files.length,
-                              ),
-                            ),
-                          ),
-                          Padding(
-                            padding: const EdgeInsets.symmetric(vertical: 16.0),
-                            child: ElevatedButton(
-                              onPressed: _selectedPath?.let((path) => () =>
-                                  context.read<ConfigurationBloc>().load(path)),
-                              child: Padding(
-                                padding: const EdgeInsets.all(8.0),
-                                child: Text(AppLocalizations.of(context)
-                                    ?.action_open_selected_file ??
-                                    'action_open_selected_file'),
-                              ),
-                            ),
-                          ),
-                          const OrSeparator(),
-                        ],
-                      ),
-                      orElse: () => const SizedBox.shrink(),
-                    ),
-                    Padding(
-                      padding: const EdgeInsets.symmetric(vertical: 16.0),
-                      child: OutlinedButton(
-                        onPressed: () => _pickFile(context),
-                        child: Padding(
-                          padding: const EdgeInsets.all(8.0),
-                          child: Text(AppLocalizations.of(context)
-                              ?.action_open_external_file ??
-                              'action_open_external_file'),
+            child: switch (state) {
+              FetchingConfigurationState() => const CircularProgressIndicator(),
+              _ => IntrinsicWidth(
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.stretch,
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: [
+                      Padding(
+                        padding: const EdgeInsets.symmetric(vertical: 16.0),
+                        child: Text(
+                          AppLocalizations.of(context)?.label_welcome ??
+                              'label_welcome',
+                          style: Theme.of(context).textTheme.displayMedium,
+                          textAlign: TextAlign.center,
                         ),
                       ),
-                    ),
-                  ],
+                      switch (state) {
+                        FetchedConfigurationState(:final files) => Column(
+                            crossAxisAlignment: CrossAxisAlignment.stretch,
+                            children: [
+                              YaruSection(
+                                margin:
+                                    const EdgeInsets.symmetric(vertical: 16.0),
+                                padding: const EdgeInsets.all(16.0),
+                                headline: Text(AppLocalizations.of(context)
+                                        ?.label_select_configuration ??
+                                    'label_select_configuration'),
+                                child: SizedBox(
+                                  height: 500.0,
+                                  width: 1000.0,
+                                  child: AlignedGridView.extent(
+                                    padding: const EdgeInsets.all(16.0),
+                                    maxCrossAxisExtent: 200.0,
+                                    mainAxisSpacing: 16.0,
+                                    crossAxisSpacing: 16.0,
+                                    itemBuilder: (context, index) => FileCard(
+                                      files[index],
+                                      selected: files[index] == _selectedPath,
+                                      onTap: () => setState(
+                                        () => _selectedPath = files[index],
+                                      ),
+                                    ),
+                                    itemCount: files.length,
+                                  ),
+                                ),
+                              ),
+                              Padding(
+                                padding:
+                                    const EdgeInsets.symmetric(vertical: 16.0),
+                                child: ElevatedButton(
+                                  onPressed: _selectedPath?.let((path) => () =>
+                                      context
+                                          .read<ConfigurationBloc>()
+                                          .load(path)),
+                                  child: Padding(
+                                    padding: const EdgeInsets.all(8.0),
+                                    child: Text(AppLocalizations.of(context)
+                                            ?.action_open_selected_file ??
+                                        'action_open_selected_file'),
+                                  ),
+                                ),
+                              ),
+                              const OrSeparator(),
+                            ],
+                          ),
+                        _ => const SizedBox.shrink(),
+                      },
+                      Padding(
+                        padding: const EdgeInsets.symmetric(vertical: 16.0),
+                        child: OutlinedButton(
+                          onPressed: () => _pickFile(context),
+                          child: Padding(
+                            padding: const EdgeInsets.all(8.0),
+                            child: Text(AppLocalizations.of(context)
+                                    ?.action_open_external_file ??
+                                'action_open_external_file'),
+                          ),
+                        ),
+                      ),
+                    ],
+                  ),
                 ),
-              ),
-            ),
+            },
           ),
         ),
       );
